+++ .changeset/neat-kings-clap.md
@@ -1,0 +1,5 @@
+---
+"@farcaster/hub-shuttle": minor
+---
+
+feat: initial release of hub-shuttle
