--- packages/hub-replicator/src/replicator/mesageHandler.ts
@@ -1,1 +1,0 @@
-export type MessageHandler = {};
