+++ packages/hub-shuttle/src/shuttle.integration.test.ts
@@ -1,19 +1,19 @@
-import { getDbClient, Tables } from "./replicator/db";
+import { getDbClient, Tables } from "./shuttle/db";
 import { migrateToLatest } from "./app/migration";
 import { log } from "./log";
 import { Kysely, sql } from "kysely";
 import { App } from "./app/app";
 import { Factories, HubEvent, HubEventType } from "@farcaster/hub-nodejs";
-import { RedisClient } from "./replicator/redis";
-import { HubSubscriber } from "./replicator/hubSubscriber";
+import { RedisClient } from "./shuttle/redis";
+import { HubSubscriber } from "./shuttle/hubSubscriber";
 import { sleep } from "./utils";
-import { MessageReconciliation } from "./replicator/messageReconciliation";
+import { MessageReconciliation } from "./shuttle/messageReconciliation";
 
 let db: Kysely<Tables>;
 let subscriber: FakeHubSubscriber;
 let redis: RedisClient;
 
-const POSTGRES_URL = process.env["POSTGRES_URL"] || "postgres://replicator:password@localhost:6541";
+const POSTGRES_URL = process.env["POSTGRES_URL"] || "postgres://shuttle:password@localhost:6541";
 const REDIS_URL = process.env["REDIS_URL"] || "localhost:16379";
 
 class FakeHubSubscriber extends HubSubscriber {
@@ -27,13 +27,13 @@ class FakeHubSubscriber extends HubSubscriber {
 }
 
 beforeAll(async () => {
-  const dbName = "replicator_test";
+  const dbName = "shuttle_test";
   if (process.env["NODE_ENV"] !== "test") {
     throw new Error("NODE_ENV must be set to test");
   }
   db = getDbClient(POSTGRES_URL);
-  await sql`DROP DATABASE IF EXISTS replicator_test`.execute(db);
-  await sql`CREATE DATABASE replicator_test`.execute(db);
+  await sql`DROP DATABASE IF EXISTS shuttle_test`.execute(db);
+  await sql`CREATE DATABASE shuttle_test`.execute(db);
 
   db = getDbClient(`${POSTGRES_URL}/${dbName}`);
   const result = await migrateToLatest(db, log);
@@ -47,7 +47,7 @@ afterAll(async () => {
   await db.destroy();
 });
 
-describe("replicator", () => {
+describe("shuttle", () => {
   let app: App;
 
   beforeEach(() => {
@@ -56,7 +56,7 @@ describe("replicator", () => {
   });
 
   test("replicates hubs events to database tables", async () => {
-    await expect(redis.getLastProcessedEvent("replicator")).resolves.toBe(0);
+    await expect(redis.getLastProcessedEvent("shuttle")).resolves.toBe(0);
     await app.start();
     const signer = Factories.Ed25519Signer.build();
     const message = await Factories.CastAddMessage.create({}, { transient: { signer } });
@@ -65,7 +65,7 @@ describe("replicator", () => {
 
     await sleep(100);
 
-    await expect(redis.getLastProcessedEvent("replicator")).resolves.toBe(10);
+    await expect(redis.getLastProcessedEvent("shuttle")).resolves.toBe(10);
     await app.stop();
 
     // count messages table and expect it to be 1
